CI: Run on `windows` machine
As this crate concerns a Windows-only API, relying on COM, it only makes
sense to lint and test it on such a machine.  The CI setup (probably
inherited as-is from our `rust-template` repository) was using an
Ubuntu machine which surprisingly allowed us to lint-test everything,
up until weird Linux-specific changes in `0.58` suddenly disable all
code inside the `windows` crate.