Extend `windows(-core)` dependency range to `0.57-0.61`
The generated code now uses `Ref` in native symbol definitions, which
is already used in all member functions; unfortunately all Windows
code uses `transmute_copy()` to construct this but it does show that we
didn't/don't really need to consume ownership of the `device: IUnknown`
(previously leaked when passed as a `device.clone()` to a native
function that doesn't "consume ownership" of the pointer).

We can remove the heavy checked-in `Windows.Win32.winmd` which has
been available behind `--in default` in `bindings.txt` for a while.

Finally, take this chance to update our `WinmdGenerator` tooling and
Windows SDK version.

https://github.com/microsoft/windows-rs/releases/tag/0.62.0
https://github.com/microsoft/windows-rs/releases/tag/63