docs.rs: Only build for Windows (and target it by default) (#40)
It seems updates to `windows` crates post `0.57` (i.e. after our `0.3.0`
release) have extra `cfg(windows)` in place that prevent building
documentation on all non-Windows platforms:

https://docs.rs/crate/amd-ext-d3d/0.3.1/builds/1291996
https://docs.rs/crate/amd-ext-d3d/0.4.0/builds/2537501

Limit the documentation to only build on a few architectures for Windows
MSVC.  Besides that, when no `default-target` is specified the first
element of `targets` is used, placing users directly on documentation
for the Windows "variant" of this crate.