ci: Fix short form of `exclude` `pathspec` in `git status --porcelain`
`!` no longer seems to properly exclude `.windows` but instead excludes
all files, allowing unexpected differences to sneak in.  Reading
https://git-scm.com/docs/gitglossary#Documentation/gitglossary.txt-pathspec
to the letter says that all special commands start with `:` followed
by a letter (`!` or `^` in this case) or the full `(exclude)` command
written out for the same meaning.  Prefixing `!` with `:` shows
the diff as expected again.